requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.8
django-filter==2.4.0
djangorestframework==3.12.1
importlib-metadata==2.0.0
Markdown==3.3.1
python-decouple==3.3
pytz==2020.1
sqlparse==0.4.1
static3==0.7.0
zipp==3.3.0
